github_export/requirements.txt
flask
flask-login
google-generativeai
python-docx
docxtpl
pandas
openpyxl
werkzeug
python-dotenv
Pillow
